refactor(assert): Unhardcode exception expectations
