+++ src/Assert.php
@@ -191,14 +191,15 @@ final class Assert
     /**
      * Fails the test.
      *
-     * Sets an expectation that the test will fail with a specific {@see AssertException} instance,
-     * then throws the exception to end the test.
-     * If the test catches the exception and continues execution, it will be marked as Risky.
+     * Use this method to explicitly indicate that the test should end in failure.
      *
-     * @param string|null $message The reason for the failure.
+     * If this method is called, it is expected that the test will end with an exception thrown by this method.
+     * If the test catches this exception and continues execution, it will be marked as Risky.
+     *
+     * @param string $message The reason for the failure.
      * @throws AssertException always, with the provided message.
      */
-    public static function fail(?string $message = null): never
+    public static function fail(string $message = ''): never
     {
         $exception = AssertException::fail($message);
         StaticState::expectException($exception);
@@ -210,7 +211,7 @@ final class Assert
      *
      * @param class-string|\Throwable $classOrObject The expected exception class, interface, or an exception object.
      *
-     * @note Requires {@see ExpectExceptionInterceptor} to be registered.
+     * @note Requires {@see ExpectationsInterceptor} to be registered.
      */
     public static function exception(
         string|\Throwable $classOrObject,
