+++ src/Assert/TestState.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace Testo\Assert;
 
-use Testo\Assert\State\ExpectedException;
 use Testo\Assert\State\Record;
+use Testo\Test\Dto\TestResult;
 
 /**
  * Collects assertions.
@@ -18,11 +18,16 @@ final class TestState
     public array $history = [];
 
     /**
-     * @var ExpectedException|null Expected exception configuration.
+     * @see Assert::leaks()
      */
-    public ?ExpectedException $expectException = null;
-
     public \WeakMap $weakMap;
+
+    public ?\Throwable $failure = null;
+
+    /**
+     * @var list<callable(TestResult, TestState): TestResult> List of expectation handlers.
+     */
+    public array $expectations = [];
 
     public function __construct()
     {
