+++ src/Module/Interceptor/InterceptorProvider.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Testo\Module\Interceptor;
 
 use Testo\Assert\Interceptor\AssertCollectorInterceptor;
-use Testo\Assert\Interceptor\ExpectExceptionInterceptor;
+use Testo\Assert\Interceptor\ExpectationsInterceptor;
 use Testo\Assert\Interceptor\ObjectTrackerInterceptor;
 use Testo\Attribute\Interceptable;
 use Testo\Common\Container;
@@ -55,7 +55,7 @@ final class InterceptorProvider
             new InstantiateTestCase(),
             new AssertCollectorInterceptor(),
             AttributesInterceptor::class,
-            new ExpectExceptionInterceptor(),
+            new ExpectationsInterceptor(),
             new ObjectTrackerInterceptor(),
         ]);
     }
